[rl] move alphabet-sort configs into examples/alphabet_sort
Each example now owns its config_registry.py (matching search_r1), discovered via
--module torchtitan.experiments.rl.examples.<name>. Move the nine alphabet-sort
GRPO configs there and drop the central rl/config_registry.py aggregator;
integration tests, README and generate.py use the full --module path now.

Pull the batch-invariance helpers into rl/batch_invariance.py: the
BatchInvariantFlexConverter (was inline in config_registry) plus
patch_bmm_for_batch_invariance / force_logprobs_fn_for_batch_invariance (were in
actors/generator.py), so any config — not just alphabet-sort — can reuse them.
