+++ torchtitan/experiments/rl/scripts/loss_compare.py
@@ -120,7 +120,7 @@ def run_training(
         "-m",
         "torchtitan.experiments.rl.train",
         "--module",
-        "rl",
+        "torchtitan.experiments.rl.examples.alphabet_sort",
         "--config",
         config,
         f"--hf_assets_path={hf_assets_path}",
