+++ torchtitan/experiments/rl/tests/test_bitwise_parity.py
@@ -59,7 +59,7 @@ from torchtitan.distributed.utils import (
     set_batch_invariance,
 )
 from torchtitan.experiments.rl.actors.trainer import compute_logprobs
-from torchtitan.experiments.rl.config_registry import (
+from torchtitan.experiments.rl.examples.alphabet_sort.config_registry import (
     rl_grpo_qwen3_0_6b_flex_batch_invariant,
     rl_grpo_qwen3_0_6b_varlen_batch_invariant,
 )
